Test Was in the Wrong Place
