+++ tests/pmv_test.py
@@ -3,11 +3,11 @@
 import unittest
 import pytest
 
-from ladybug.comfort.pmv import PMVComfortParameters
+from ladybug.comfort.standard.pmv import PMVComfortParameters
 
-from ladybug.comfort.utility.pmv import pmv
-from ladybug.comfort.utility.pmv import pmv_fanger
-from ladybug.comfort.utility.pmv import pierce_set
+from ladybug.comfort.pmv import pmv
+from ladybug.comfort.pmv import pmv_fanger
+from ladybug.comfort.pmv import pierce_set
 
 
 class PMVTestCase(unittest.TestCase):
@@ -23,22 +23,19 @@ class PMVTestCase(unittest.TestCase):
 
     def test_pmv_fanger_utility(self):
         """Test the pmv_fanger utility function"""
-        pmv_comf, ppd = pmv_fanger(19, 23, 0.5, 60, 1.5, 0.4)
-
-        assert pmv_comf == pytest.approx(-1.74, rel=1e-2)
-        assert round(ppd) == 64
+        pmv_comf, ppd, hl = pmv_fanger(19, 23, 0.1, 60, 1.5, 0.4)
+        assert pmv_comf == pytest.approx(-0.680633, rel=1e-2)
+        assert ppd == pytest.approx(14.7373, rel=1e-2)
 
     def test_pierce_set_utility(self):
         set = pierce_set(19, 23, 0.5, 60, 1.5, 0.4)
-
-        assert set == pytest.approx(-18.7, rel=1e-1)
+        assert set == pytest.approx(18.8911, rel=1e-2)
 
     def test_pmv_utility(self):
         result = pmv(19, 23, 0.5, 60, 1.5, 0.4)
-
-        assert result['pmv'] == pytest.approx(-1.74, rel=1e-2)
-        assert round(result['ppd']) == 64
-        assert result['set'] == pytest.approx(-18.7, rel=1e-1)
+        assert result['pmv'] == pytest.approx(-1.6745, rel=1e-2)
+        assert round(result['ppd']) == pytest.approx(60.382974, rel=1e-2)
+        assert result['set'] == pytest.approx(18.8911, rel=1e-2)
 
     def test_pmv_comfort_parameters(self):
         """Test PMVComfortParameters."""
@@ -51,6 +48,34 @@ class PMVTestCase(unittest.TestCase):
             ppd_comfort_thresh, humid_ratio_up, humid_ratio_low, still_air_thresh)
 
         assert pmv_comf.ppd_comfort_thresh == ppd_comfort_thresh
-        assert pmv_comf.humid_ratio_up == humid_ratio_up
-        assert pmv_comf.humid_ratio_low == humid_ratio_low
+        assert pmv_comf.humid_ratio_upper == humid_ratio_up
+        assert pmv_comf.humid_ratio_lower == humid_ratio_low
         assert pmv_comf.still_air_threshold == still_air_thresh
+
+    def test_comfort_check(self):
+        """Test comfort check on PMVComfortParameters."""
+        pmv_comf = PMVComfortParameters()
+        comf_test = pmv_comf.is_comfortable(13, 0.01)
+        assert comf_test is False
+        comf_test = pmv_comf.is_comfortable(7, 0.01)
+        assert comf_test is True
+
+    def test_thermal_condition_check(self):
+        """Test the thermal condition check on PMVComfortParameters."""
+        pmv_comf = PMVComfortParameters()
+        condition_test = pmv_comf.thermal_condition(-1, 20)
+        assert condition_test == -1
+        condition_test = pmv_comf.thermal_condition(0, 5)
+        assert condition_test == 0
+
+    def test_discomfort_reason_check(self):
+        """Test the thermal condition check on PMVComfortParameters."""
+        pmv_comf = PMVComfortParameters()
+        condition_test = pmv_comf.discomfort_reason(-1, 20, 0.01)
+        assert condition_test == -1
+        condition_test = pmv_comf.discomfort_reason(0, 5, 0.01)
+        assert condition_test == 0
+
+
+if __name__ == "__main__":
+    unittest.main()
